+++ src/Contracts/Api/OrderApiInterface.php
@@ -8,6 +8,7 @@ use MasyaSmv\FinamSdk\Dto\Order\CancelOrderRequest;
 use MasyaSmv\FinamSdk\Dto\Order\OrderRequest;
 use MasyaSmv\FinamSdk\Dto\Order\OrdersRequest;
 use MasyaSmv\FinamSdk\Dto\Order\PlaceOrderRequest;
+use MasyaSmv\FinamSdk\Dto\Order\PlaceSlTpOrderRequest;
 use MasyaSmv\FinamSdk\Dto\Transport\ApiResponse;
 
 interface OrderApiInterface
@@ -17,6 +18,8 @@ interface OrderApiInterface
     public function order(OrderRequest $request): ApiResponse;
 
     public function place(PlaceOrderRequest $request): ApiResponse;
+
+    public function placeSlTp(PlaceSlTpOrderRequest $request): ApiResponse;
 
     public function cancel(CancelOrderRequest $request): ApiResponse;
 }