+++ src/Contracts/Session/SessionOrderServiceInterface.php
@@ -7,6 +7,7 @@ namespace MasyaSmv\FinamSdk\Contracts\Session;
 use MasyaSmv\FinamSdk\Collections\OrderCollection;
 use MasyaSmv\FinamSdk\Dto\Order\OrderDto;
 use MasyaSmv\FinamSdk\Dto\Order\PlaceOrderInputDto;
+use MasyaSmv\FinamSdk\Dto\Order\PlaceSlTpOrderInputDto;
 
 interface SessionOrderServiceInterface
 {
@@ -15,4 +16,6 @@ interface SessionOrderServiceInterface
     public function getOrder(string $orderId, ?string $accountId = null): OrderDto;
 
     public function placeOrder(PlaceOrderInputDto $order, ?string $accountId = null): OrderDto;
+
+    public function placeSlTpOrder(PlaceSlTpOrderInputDto $order, ?string $accountId = null): OrderDto;
 }